- cleaned up some code - added lighting spreading though chunks - improvements to fluids

+++ Minecraft Clone/src/Core/Interfaces/ClientInterface.h
@@ -19,7 +19,7 @@ public:
 
     ClientInterface() = default;
 
-    int PollClientPlayerAction(std::vector<Packet::PlayerAction>& playerAction) {
+    size_t PollClientPlayerAction(std::vector<Packet::PlayerAction>& playerAction) {
         size_t prevSize = playerAction.size();
         player_action_queue_.PopAll(playerAction);
         return playerAction.size() - prevSize;
